Rename spacecraft directory
