+++ s2e-ff/src/simulation/spacecraft/FfComponents.hpp
