+++ s2e-ff/src/simulation/spacecraft/FfComponents2.cpp
