+++ s2e-ff/src/simulation/spacecraft/FfSat.hpp
